+++ cloudsim-plus/src/main/java/org/cloudbus/cloudsim/hosts/HostNull.java
@@ -62,7 +62,7 @@ final class HostNull implements Host {
     @Override public int getNumberOfFreePes() {
         return 0;
     }
-    @Override public int getNumberOfPes() {
+    @Override public long getNumberOfPes() {
         return 0;
     }
     @Override public double getMips() { return 0; }
