+++ cloudsim-plus/src/main/java/org/cloudbus/cloudsim/hosts/power/PowerHostNull.java
@@ -112,7 +112,7 @@ final class PowerHostNull implements PowerHost {
     @Override public int getNumberOfFreePes() {
         return 0;
     }
-    @Override public int getNumberOfPes() {
+    @Override public long getNumberOfPes() {
         return 0;
     }
     @Override public double getMips() { return 0; }
